Added game controller, service, dto and changed every money related field in the entities from int64 to float64

+++ models/Game.entity.go
@@ -29,9 +29,9 @@ type Game struct {
 	Description string         `json:"description" gorm:"type:text"`
 	Category    GameCategory   `json:"category" gorm:"type:varchar(30);not null"`
 	Status      GameStatus     `json:"status" gorm:"type:varchar(20);default:'active'"`
-	MinBet      int64          `json:"min_bet" gorm:"not null"`
-	MaxBet      int64          `json:"max_bet" gorm:"not null"`
-	HouseEdge   int64          `json:"house_edge" gorm:"default:0"`
+	MinBet      float64        `json:"min_bet" gorm:"not null"`
+	MaxBet      float64        `json:"max_bet" gorm:"not null"`
+	HouseEdge   float64        `json:"house_edge" gorm:"default:0"`
 	CreatedAt   time.Time      `json:"created_at" gorm:"autoCreateTime"`
 	UpdatedAt   time.Time      `json:"updated_at" gorm:"autoUpdateTime"`
 	DeletedAt   gorm.DeletedAt `json:"-" gorm:"index"`
